클라 코드 수정
씬 전환을 위해 씬에서 처리해야 될 일을 프레임워크에서 처리하지 않게 코드를 정리함.

- 환경맵, 미니맵같은 경우에는 로비 씬에서는 필요 없으므로 배틀 씬에서만 쓰도록 관련 코드들을 씬 클래스로 옮김.

- 그로 인해 렌더 타겟 뷰, 깊이 스텐실 뷰 서술자 힙을 프레임워크에서 분리시킴. 원래는 프레임워크에서 만든 서술자 힙에 갯수를 추가해서 같이 썼었는데 그렇게 하지 않고 씬에서 아예 새로 만들어서 사용하도록 변경함.

+++ Project Client/PillsFighter/Font.cpp
@@ -190,16 +190,16 @@ void CFont::LoadDataFromFile(const char *pstrFileName)
 				else if (!strncmp(pstrToken, "pad", 3)) // padding=
 				{
 					char pstrValue = pstrToken[8];
-					m_fTopPadding = atof(&pstrValue) / FRAME_BUFFER_WIDTH;
+					m_fTopPadding = (float)(atof(&pstrValue) / FRAME_BUFFER_WIDTH);
 
 					pstrValue = pstrToken[10];
-					m_fRightPadding = atof(&pstrValue) / FRAME_BUFFER_WIDTH;
+					m_fRightPadding = (float)(atof(&pstrValue) / FRAME_BUFFER_WIDTH);
 
 					pstrValue = pstrToken[12];
-					m_fBottomPadding = atof(&pstrValue) / FRAME_BUFFER_WIDTH;
+					m_fBottomPadding = (float)(atof(&pstrValue) / FRAME_BUFFER_WIDTH);
 
 					pstrValue = pstrToken[14];
-					m_fLeftPadding = atof(&pstrValue) / FRAME_BUFFER_WIDTH;
+					m_fLeftPadding = (float)(atof(&pstrValue) / FRAME_BUFFER_WIDTH);
 				}
 				else if (!strncmp(pstrToken, "spa", 3)) // spacing=
 				{
@@ -218,13 +218,13 @@ void CFont::LoadDataFromFile(const char *pstrFileName)
 				{
 					char *pstr = &pstrToken[11];
 
-					m_fLineHeight = atof(pstr) / FRAME_BUFFER_HEIGHT;
+					m_fLineHeight = (float)(atof(pstr) / FRAME_BUFFER_HEIGHT);
 				}
 				else if (!strncmp(pstrToken, "base=", 5)) // base=
 				{
 					char *pstr = &pstrToken[5];
 
-					m_fBaseHeight = atof(pstr) / FRAME_BUFFER_HEIGHT;
+					m_fBaseHeight = (float)(atof(pstr) / FRAME_BUFFER_HEIGHT);
 				}
 				else if (!strncmp(pstrToken, "scaleW", 6)) // scaleW=
 				{
@@ -422,7 +422,7 @@ CTextObject* CFont::SetText(const char *pstrText, XMFLOAT2 xmf2Position, XMFLOAT
 
 	XMFLOAT2 xmf2ChPosition = xmf2Position;
 
-	int nLength = strlen(pstrText);
+	int nLength = (int)strlen(pstrText);
 	CFontVertex *pFontVertices = new CFontVertex[nLength];
 
 	for (int i = 0; i < nLength; i++)
